rffds

+++ src/processor/cancel.java
@@ -21,23 +21,27 @@ public class cancel extends AbstractProcessor{
 		String head = data[0] + Server.token;
 		boolean isSuccess = false;
 		String rs = "";
+		double balance = 0;
 		try{
 		if( CommonMethods.checkPasswd(data[2], data[3]) && CommonMethods.checkUid(data[1],data[2]) )
 		{
-			String SQL_1 = "SELECT TYPE FROM SYSTEM.USER, SYSTEM.ACCOUNT WHERE " +
+			String SQL_1 = "SELECT USER.TYPE,BALANCE FROM SYSTEM.USER, SYSTEM.ACCOUNT WHERE " +
 					"ACCOUNT.UID=USER.UID AND ACCOUNT.AID='" + data[2] + "';";
 			ResultSet rss = QueryProcessor.query(SQL_1);
 			if( rss.next() )
 			{
-				String type = rss.getString("TYPE");
+				String type = rss.getString("USER.TYPE");
+				balance = rss.getDouble("BALANCE");
 				String SQL_del = "DELETE FROM SYSTEM.ACCOUNT WHERE AID='" + data[2] + "';";
-				if( UpdateProcessor.update(SQL_del) != 0 ){
+				if( balance >=0 && UpdateProcessor.update(SQL_del) != 0 ){
 					if( type.equals("e") )
 					{
 						SQL_del = "DELETE FROM SYSTEM.OPERATOR WHERE AID='" + data[2] + "';";
 						if( UpdateProcessor.update(SQL_del) != 0 )
 							isSuccess = true;
 					}
+					else
+						isSuccess = true;
 				}
 			}
 		}
@@ -48,11 +52,11 @@ public class cancel extends AbstractProcessor{
 		}
 		if( isSuccess )
 		{
-			rs = "successed";
-			Log.log(rd.getJobNumber(),data[2],data[0],rs);
+			rs = "success";
+			Log.log(rd.getJobNumber(),data[2],data[0],"", 0, balance, balance);
 		}
 		else
-			rs = "failed";
+			rs = "fail";
 		SendDataList.getInstance().add(
 				new SendData( rd, head + rs ));
 	}
